requirements.txt
streamlit>=1.26.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
matplotlib>=3.7.0
pillow>=9.5.0 